refactor: changed the way the fontend views and components were factor adding structure
