+++ revealJS/js/config.js
@@ -18,7 +18,7 @@ export default {
 	// Display presentation control arrows
 	controls: true,
 
-	// Help the user learn the controls by providing hints, for example by
+	// help the user learn the controls by providing hints, for example by
 	// bouncing the down arrow when they first encounter a vertical slide
 	controlsTutorial: true,
