+++ src/frontend/src/components/home/NavigationBar.vue
