+++ src/frontend/src/components/home/clients/__tests__/ClientDesk.cy.ts
@@ -1,6 +1,6 @@
 import { createVuetify } from 'vuetify'
 import { mount } from 'cypress/vue'
-import ClientDesk from '@/views/ClientDesk.vue'
+import ClientDesk from 'WorkWise-Central/src/frontend/src/views/ClientDesk.vue'
 
 it('renders client desk', () => {
   const vuetify = createVuetify()
