+++ src/frontend/src/components/home/clients/__tests__/ClientDetail.cy.ts
@@ -1,27 +1,21 @@
 import { createVuetify } from 'vuetify'
 import { mount } from 'cypress/vue'
-import JoinCompanyModal from './JoinCompanyModal.vue'
+import ClientDetails from '../management/AddClient.vue'
 
 it('renders client detials', () => {
   const vuetify = createVuetify()
-  mount(JoinCompanyModal, {
+  mount(ClientDetails, {
     global: {
       plugins: [vuetify]
     }
   })
-  cy.contains('Join Company')
-  cy.contains('Search for the company by name')
-  cy.contains('OR')
-  cy.contains('Enter the Company ID')
 })
 
 it('click create client button', () => {
   const vuetify = createVuetify()
-  mount(JoinCompanyModal, {
+  mount(ClientDetails, {
     global: {
       plugins: [vuetify]
     }
   })
-
-  cy.contains('button', 'JOIN COMPANY').click()
 })
