+++ src/frontend/src/components/home/clients/management/ClientDetails.vue
