+++ src/frontend/src/components/home/clients/management/DeleteClient.vue
