+++ src/frontend/src/components/home/dashboard/home/DashboardScreen.vue
