+++ src/frontend/src/components/home/jobs/Kanban.vue
