+++ src/frontend/src/components/home/jobs/__tests__/AddJob.cy.ts
@@ -1,6 +1,6 @@
 import { createVuetify } from 'vuetify'
 import { mount } from 'cypress/vue'
-import JobDetails from './AddJob.vue'
+import JobDetails from '../management/AddJob.vue'
 
 it('renders job detials', () => {
   const vuetify = createVuetify()
