+++ src/frontend/src/components/home/jobs/management/JobCard.vue
