+++ src/frontend/src/components/home/jobs/management/JobChecklist.vue
