+++ src/frontend/src/components/home/jobs/management/JobLabels.vue
