+++ src/frontend/src/components/signup/JoinCompanyModal.vue
