+++ src/frontend/src/components/signup/RegisterCompanyModal.vue
