+++ src/frontend/src/components/signup/SplashScreen.vue
