+++ src/frontend/src/views/ClientEmployee.vue
@@ -4,7 +4,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import NavigationBar from '@/components/NavigationBar.vue'
+import NavigationBar from '@/components/home/NavigationBar.vue'
 import ClientDesk from '@/components/ClientEmployeeView.vue'
 export default defineComponent({
   name: 'EmployeeDesk',
