+++ src/frontend/src/views/EmployeeDesk.vue
@@ -4,8 +4,8 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import NavigationBar from '@/components/NavigationBar.vue'
-import ClientDesk from '@/components/ClientDesk.vue'
+import NavigationBar from '@/components/home/NavigationBar.vue'
+import ClientDesk from '@/components/home/clients/management/ClientDesk.vue'
 export default defineComponent({
   name: 'EmployeeDesk',
   computed: {
