+++ src/frontend/src/views/JobAssignment.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import NavigationBar from '../components/NavigationBar.vue'
+import NavigationBar from '../components/home/NavigationBar.vue'
 </script>
 
 <template>
