+++ src/frontend/src/views/ManagerView.vue
@@ -7,8 +7,8 @@
 
 <script>
 import { defineComponent } from 'vue'
-import NavigationBar from '../components/NavigationBar.vue'
-import ClientDesk from '../components/ClientDesk.vue'
+import NavigationBar from '../components/home/NavigationBar.vue'
+import ClientDesk from '../components/home/clients/management/ClientDesk.vue'
 export default defineComponent({
   name: 'ManagerView',
   components: {
