+++ src/frontend/src/views/emailConfirmation/EmailConfirmation.vue
