+++ src/frontend/src/views/home/dashboard/DashboardView.vue
@@ -4,8 +4,8 @@
 
 <script>
 import { defineComponent } from 'vue'
-import DashboardScreen from '@/components/DashboardScreen.vue'
-import NavigationBar from '@/components/NavigationBar.vue'
+import DashboardScreen from '@/components/home/dashboard/home/DashboardScreen.vue'
+import NavigationBar from '@/components/home/NavigationBar.vue'
 export default defineComponent({
   name: 'DashboardView',
   computed: {
