+++ src/frontend/src/views/home/employees/ManageEmployeesT.vue
@@ -4,8 +4,8 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import ManageEmployeesT from '@/components/ManageEmployeesT.vue'
-import NavigationBar from '@/components/NavigationBar.vue'
+import ManageEmployeesT from '@/components/home/employees/management/ManageEmployeesT.vue'
+import NavigationBar from '@/components/home/NavigationBar.vue'
 export default defineComponent({
   name: 'ManagerEmployeesView',
   computed: {
