+++ src/frontend/src/views/home/employees/ManagerEditEmployee.vue
@@ -4,8 +4,8 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import ManagerEditEmployee from '@/components/ManagerEditEmployee.vue'
-import NavigationBar from '@/components/NavigationBar.vue'
+import ManagerEditEmployee from '@/components/home/employees/management/ManagerEditEmployee.vue'
+import NavigationBar from '@/components/home/NavigationBar.vue'
 export default defineComponent({
   name: 'ManagerEditEmployeeView',
   computed: {
