+++ src/frontend/src/views/home/help/Support.vue
@@ -4,8 +4,8 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import Support from '@/components/Support.vue'
-import NavigationBar from '@/components/NavigationBar.vue'
+import Support from '@/components/home/help/support/Support.vue'
+import NavigationBar from '@/components/home/NavigationBar.vue'
 export default defineComponent({
   name: 'SupportView',
   computed: {
